Update on build process
