+++ .eslintignore
@@ -1,3 +1,2 @@
-config/**
 dist/**
 node_modules/**
