+++ src/elements_data.js
@@ -1,3 +1,5 @@
+import C from './constants.js';
+
 const dicomNDict = {
   0x00020000: { vr: 'UL',
     vm: C.VM_SINGLE,
