+++ src/version.js
@@ -1,0 +1,1 @@
+export default '0.0.1';
